Style changes in footer, fixes permission denied errors for non-logged in users for audio streaming.

+++ app/helpers/content_item_helper.rb
@@ -57,7 +57,8 @@ module ContentItemHelper
 
     # Allow user to clone this item if they are able to act as 
     # contributors, and if this item isn't already a clone.
-    if current_user.can_act_as?("contributor") && 
+    if logged_in? && 
+        current_user.can_act_as?("contributor") && 
         content_item.parent_id.nil?
       links << link_to( 'create clone', content_item_clones_path(content_item),
                         :confirm => "Are you sure that you wish  to create " + 
@@ -68,13 +69,6 @@ module ContentItemHelper
 
     if content_item.writable_by?(current_user)
       links << link_to("edit", edit_content_item_path(content_item))
-      links << link_to("images", 
-                       images_path(:imageable_type => "content_item", 
-                                   :imageable_id => content_item.id))
-
-      links << link_to("audio", 
-                       audio_files_path(:audible_type => "content_item", 
-                                       :audible_id => content_item.id))
 
       links << link_to("versions", 
                        content_item_versions_path(content_item))
@@ -83,12 +77,23 @@ module ContentItemHelper
 
     end
 
-    if current_user.can_act_as?("editor") &&
+    links << link_to("images", 
+                     images_path(:imageable_type => "content_item", 
+                                 :imageable_id => content_item.id))
+
+    links << link_to("audio", 
+                     audio_files_path(:audible_type => "content_item", 
+                                      :audible_id => content_item.id))
+
+
+    if logged_in? &&
+        current_user.can_act_as?("editor") &&
         !content_item.private_clones.empty?
       links << link_to('view clones', content_item_clones_path(content_item))
     end
 
-    if current_user.can_act_as?("administrator")
+    if logged_in? && 
+        current_user.can_act_as?("administrator")
       links << link_to('delete', content_item_path(content_item),
                        :confirm => "Are you sure?",
                        :method => :delete)
